Create a list for the event and add attendees as members of the list.
Note that calling Twitter.list_create for a list that already exists doesn't create a duplicate list nor raise an exception. And calling Twitter.list_add_members for members that are already in the list doesn't add duplicate members nor raise an exception.

Note also you will need to add the real credentials in a config.yml file in the same format as config.example.yml.